dashboard/requirements.txt
streamlit>=1.30
plotly>=5.18
pandas>=2.1
psycopg2-binary>=2.9
python-dotenv>=1.0
